+++ backend/src/main/java/ruan/eloy/backend/Exceptions/StudentNotFoundException.java
@@ -1,8 +1,7 @@
 package ruan.eloy.backend.Exceptions;
 
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
-public class StudentNotFoundException extends UsernameNotFoundException {
+public class StudentNotFoundException extends RuntimeException {
 
     public StudentNotFoundException(String msg) {
         super("Student not found. " + msg);
@@ -12,7 +11,4 @@ public class StudentNotFoundException extends UsernameNotFoundException {
         super("Student not found");
     }
 
-    public StudentNotFoundException(String msg, Throwable t) {
-        super("Student not found. " + msg, t);
-    }
 }
