+++ backend/src/main/java/ruan/eloy/backend/dto/SignUpRequest.java
@@ -1,4 +1,4 @@
-package ruan.eloy.backend.payload;
+package ruan.eloy.backend.dtos;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
